make torch package
