+++ python/ray/ml/integrations/torch/__init__.py
@@ -1,0 +1,3 @@
+from ray.ml.integrations.torch.torch_predictor import TorchPredictor
+
+__all__ = ["TorchPredictor"]
